refactor: rename plugin-interfaces to plugin-types
