+++ src/plugin-types.ts
