+++ src/span-data.ts
@@ -17,7 +17,7 @@
 'use strict';
 
 import { Constants } from './constants';
-import { SpanData as SpanDataInterface } from './plugin-interfaces';
+import { SpanData as SpanDataInterface } from './plugin-types';
 import { Trace } from './trace';
 import { TraceLabels } from './trace-labels';
 import { TraceSpan } from './trace-span';
